Replace the `--command-channel` flag with `--enable-dm`.
This reworks Lurch's group interactions. By default playbook commands can only
be run in a channel or group so that everyone is kept abreast of actions;
attempting to run playbooks via direct messaging will not work (although all
other commands will work). This means there is no concept of a 'command channel'
any more; any member of a group or channel of which Lurch is also a member can
issue commands to him.  Commands can also be issued via DM if Lurch is started
with the `--enable-dm` option.